+++ Tasks/UnityGetProjectVersion/UnityGetProjectVersionV1/CHANGELOG.md
@@ -5,12 +5,33 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.1]
+## [1.0.22]
+
+### Added
+
+- Added log output while running the task
 
 ### Changed
 
-- Fixed issue where project version would not correctly be determined if the ProjectVersion.txt contains a revision editor version.
+- Performance and optimization changes
+
+## [1.0.4]
+
+### Added
+
+- Added debugging messages (set debug flag to see in log)
+
+### Changed
+
+- Changed project version detection code
+- Updated node packages
+
+## [1.0.1]
+
+### Fixed
+
+- Fixed issue where project version would not correctly be determined if the ProjectVersion.txt contains a revision editor version
 
 ## [1.0.0]
 
-Initial release.
+Initial release